Roles and Permissions Added

+++ app/Http/Controllers/Admin/PermissionController.php
@@ -21,15 +21,24 @@ class PermissionController extends Controller
     }
     public function index()
     {
+        if (!auth()->user()->hasPermission('Permission Listing')) {
+            return redirect()->route('dashboard')->with('error', 'You do not have permission to view Permission Listing!');
+        }
         $permissions = $this->permissionRepository->getAllPermissions();
         return view('admin.permissions.permission', compact('permissions'));
     }
     public function create()
     {
+        if (!auth()->user()->hasPermission('Permission Add')) {
+            return redirect()->route('permissions.index')->with('error', 'You do not have permission to Add Permission!');
+        }
         return view('admin.permissions.create');
     }
     public function store(Request $request)
     {
+        if (!auth()->user()->hasPermission('Permission Add')) {
+            return redirect()->route('permissions.index')->with('error', 'You do not have permission to Add Permission!');
+        }
         $validator = Validator::make($request->all(),[
             'name'=>'required|string|max:255',
             'module'=>'required|string|max:255',
@@ -44,11 +53,17 @@ class PermissionController extends Controller
     }
     public function edit($id)
     {
+        if (!auth()->user()->hasPermission('Permission Edit')) {
+            return redirect()->route('permissions.index')->with('error', 'You do not have permission to Edit Permission!');
+        }
         $permission = $this->permissionRepository->getPermissionById($id);
         return view('admin.permissions.edit',compact('permission'));
     }
     public function update($id, Request $request)
     {
+        if (!auth()->user()->hasPermission('Permission Edit')) {
+            return redirect()->route('permissions.index')->with('error', 'You do not have permission to Edit Permission!');
+        }
         $validatedData = $request->validate([
             'name' => 'required|string|max:255',
             'module' => 'required|string|max:255',
@@ -58,11 +73,17 @@ class PermissionController extends Controller
     }
     public function show($id)
     {
+        if (!auth()->user()->hasPermission('Permission Detail')) {
+            return redirect()->route('permissions.index')->with('error', 'You do not have permission to view Permission Details!');
+        }
         $permission = $this->permissionRepository->getPermissionById($id);
         return view('admin.permissions.show',compact('permission'));
     }
     public function destroy($id)
     {
+        if (!auth()->user()->hasPermission('Permission Delete')) {
+            return redirect()->route('permissions.index')->with('error', 'You do not have permission to Delete Permission!');
+        }
         $permission = $this->permissionRepository->deletePermission($id);
         if(!$permission)
         {
@@ -72,6 +93,9 @@ class PermissionController extends Controller
     }
     public function updateStatus($id, $status)
     {
+        if (!auth()->user()->hasPermission('Permission Change Status')) {
+            return redirect()->route('permissions.index')->with('error', 'You do not have permission to Change Permission Status!');
+        }
         $permission = $this->permissionRepository->updatePermissionStatus($id, $status);
         return redirect()->route('permissions.index')->with('success', $status == 1 ? 'Permission activated successfully.':'Permission deactivated successfully.');
     }
