+++ app/Http/Controllers/Admin/RoleController.php
@@ -20,16 +20,25 @@ class RoleController extends Controller
     
     public function index()
     {
+        if (!auth()->user()->hasPermission('Role Listing')) {
+            return redirect()->route('dashboard')->with('error', 'You do not have permission to view Role Listing!');
+        }
         $roles = $this->roleRepository->getAllRoles();
         return view('admin.roles.role',compact('roles'));
     }
     public function create()
     {
+        if (!auth()->user()->hasPermission('Role Add')) {
+            return redirect()->route('roles.index')->with('error', 'You do not have permission to Add Role!');
+        }
         $permissions = $this->permissionRepository->getPermissionsByModules();
         return view('admin.roles.create', compact('permissions'));
     }
     public function store(Request $request)
     {
+        if (!auth()->user()->hasPermission('Role Add')) {
+            return redirect()->route('roles.index')->with('error', 'You do not have permission to Add Role!');
+        }
         $validator = Validator::make($request->all(), [
             'name' => 'required|string|max:255|unique:roles',
             'permissions' => 'array', 
@@ -46,17 +55,26 @@ class RoleController extends Controller
     }
     public function show($id)
     {
+        if (!auth()->user()->hasPermission('Role Detail')) {
+            return redirect()->route('roles.index')->with('error', 'You do not have permission to view Role Details!');
+        }   
         $role = $this->roleRepository->getRoleById($id);
         return view('admin.roles.show',compact('role'));
     }
     public function edit($id)
     {
+        if (!auth()->user()->hasPermission('Role Edit')) {
+            return redirect()->route('roles.index')->with('error', 'You do not have permission to Edit Role!');
+        }
         $role = $this->roleRepository->getRoleById($id);
         $permissions = $this->permissionRepository->getPermissionsByModules();
         return view('admin.roles.edit', compact('role', 'permissions'));
     }
     public function update(Request $request, $id)
     {
+        if (!auth()->user()->hasPermission('Role Edit')) {
+            return redirect()->route('roles.index')->with('error', 'You do not have permission to Edit Role!');
+        }
         $validator = Validator::make($request->all(), [
             'name' => 'required|string|max:255|unique:roles,name,' . $id,
             'permissions' => 'array',
@@ -73,6 +91,9 @@ class RoleController extends Controller
     }
     public function destroy($id)
     {
+        if (!auth()->user()->hasPermission('Role Delete')) {
+            return redirect()->route('roles.index')->with('error', 'You do not have permission to Delete Role!');
+        }
         $role = $this->roleRepository->deleteRole($id);
         if(!$role)
         {
@@ -82,6 +103,9 @@ class RoleController extends Controller
     }
     public function updateStatus($id,$status)
     {
+        if (!auth()->user()->hasPermission('Role Change Status')) {
+            return redirect()->route('roles.index')->with('error', 'You do not have permission to Change Role Status!');
+        }
         $role = $this->roleRepository->updateRoleStatus($id, $status);
         return redirect()->route('roles.index')->with('success', $status == 1 ? 'Role activated successfully.':'Role deactivated successfully.');
     }
